+++ authpass/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed possible error dialogs #124 ("The method 'markNeedsBuild' was called on null")
 * Fixed error in group list when adding new subgroups while a filter is active.
 * UI: Progress overlay on initial page now matches screen height.
+* UI: allow displaying of plain text master password during unlocking. #86
 
 # 1.7.2